+++ extensions/ql-vscode/src/common/logging/index.ts
@@ -1,5 +1,6 @@
 export * from "./logger";
 export * from "./notification-logger";
+export * from "./notifications";
 export * from "./tee-logger";
 export * from "./vscode/loggers";
 export * from "./vscode/output-channel-logger";
