+++ extensions/ql-vscode/src/compare/compare-view.ts
@@ -6,6 +6,7 @@ import {
   QueryCompareResult,
 } from "../pure/interface-types";
 import { extLogger, Logger } from "../common";
+import { showAndLogExceptionWithTelemetry } from "../common/vscode/logging";
 import { CodeQLCliServer } from "../codeql-cli/cli";
 import { DatabaseManager } from "../databases/local-databases";
 import { jumpToLocation } from "../databases/local-databases/locations";
@@ -24,8 +25,6 @@ import {
 } from "../common/vscode/abstract-webview";
 import { telemetryListener } from "../telemetry";
 import { redactableError } from "../pure/errors";
-
-import { showAndLogExceptionWithTelemetry } from "../common/vscode/log";
 
 interface ComparePair {
   from: CompletedLocalQueryInfo;
