+++ extensions/ql-vscode/src/databases/database-fetcher.ts
@@ -31,7 +31,7 @@ import {
 import { Credentials } from "../common/authentication";
 import { AppCommandManager } from "../common/commands";
 import { ALLOW_HTTP_SETTING } from "../config";
-import { showAndLogInformationMessage } from "../common/vscode/log";
+import { showAndLogInformationMessage } from "../common/logging";
 
 /**
  * Prompts a user to fetch a database from a remote location. Database is assumed to be an archive file.
