+++ extensions/ql-vscode/src/debugger/debug-configuration.ts
@@ -9,7 +9,7 @@ import { LocalQueries } from "../local-queries";
 import { getQuickEvalContext, validateQueryPath } from "../run-queries-shared";
 import * as CodeQLProtocol from "./debug-protocol";
 import { getErrorMessage } from "../pure/helpers-pure";
-import { showAndLogErrorMessage } from "../common/vscode/log";
+import { showAndLogErrorMessage } from "../common/logging";
 import { extLogger } from "../common";
 
 /**
