+++ extensions/ql-vscode/src/language-support/ast-viewer/ast-viewer.ts
@@ -27,8 +27,8 @@ import { DisposableObject } from "../../pure/disposable-object";
 import { asError, getErrorMessage } from "../../pure/helpers-pure";
 import { redactableError } from "../../pure/errors";
 import { AstViewerCommands } from "../../common/commands";
-import { showAndLogExceptionWithTelemetry } from "../../common/vscode/log";
 import { extLogger } from "../../common";
+import { showAndLogExceptionWithTelemetry } from "../../common/vscode/logging";
 
 export interface AstItem {
   id: BqrsId;
