+++ extensions/ql-vscode/src/query-evaluation-logging/eval-log-viewer.ts
@@ -12,8 +12,8 @@ import { DisposableObject } from "../pure/disposable-object";
 import { asError, getErrorMessage } from "../pure/helpers-pure";
 import { redactableError } from "../pure/errors";
 import { EvalLogViewerCommands } from "../common/commands";
-import { showAndLogExceptionWithTelemetry } from "../common/vscode/log";
 import { extLogger } from "../common";
+import { showAndLogExceptionWithTelemetry } from "../common/vscode/logging";
 
 export interface EvalLogTreeItem {
   label?: string;
