+++ extensions/ql-vscode/src/query-history/query-history-manager.ts
@@ -60,7 +60,7 @@ import {
   showAndLogErrorMessage,
   showAndLogInformationMessage,
   showAndLogWarningMessage,
-} from "../common/vscode/log";
+} from "../common/logging";
 
 /**
  * query-history-manager.ts
