+++ extensions/ql-vscode/src/query-server/legacy/run-queries.ts
@@ -13,6 +13,7 @@ import { tmpDir } from "../../tmp-dir";
 import { ProgressCallback } from "../../common/vscode/progress";
 import { QueryMetadata } from "../../pure/interface-types";
 import { extLogger, Logger } from "../../common";
+import { showAndLogExceptionWithTelemetry } from "../../common/vscode/logging";
 import * as messages from "../../pure/legacy-messages";
 import * as newMessages from "../../pure/new-messages";
 import * as qsClient from "./query-server-client";
@@ -22,10 +23,7 @@ import { QueryEvaluationInfo, QueryOutputDir } from "../../run-queries-shared";
 import { redactableError } from "../../pure/errors";
 import { CoreQueryResults, CoreQueryTarget } from "../query-runner";
 import { Position } from "../../pure/messages-shared";
-import {
-  showAndLogExceptionWithTelemetry,
-  showAndLogWarningMessage,
-} from "../../common/vscode/log";
+import { showAndLogWarningMessage } from "../../common/logging";
 import { ensureDirSync } from "fs-extra";
 
 const upgradesTmpDir = join(tmpDir.name, "upgrades");
