+++ extensions/ql-vscode/src/query-server/legacy/upgrades.ts
@@ -6,6 +6,7 @@ import {
   UserCancellationException,
 } from "../../common/vscode/progress";
 import { extLogger } from "../../common";
+import { showAndLogExceptionWithTelemetry } from "../../common/vscode/logging";
 import * as messages from "../../pure/legacy-messages";
 import * as qsClient from "./query-server-client";
 import * as tmp from "tmp-promise";
@@ -13,7 +14,6 @@ import { dirname } from "path";
 import { DatabaseItem } from "../../databases/local-databases";
 import { asError, getErrorMessage } from "../../pure/helpers-pure";
 import { redactableError } from "../../pure/errors";
-import { showAndLogExceptionWithTelemetry } from "../../common/vscode/log";
 
 /**
  * Maximum number of lines to include from database upgrade message,
