+++ extensions/ql-vscode/test/vscode-tests/no-workspace/data-extensions-editor/external-api-usage-query.test.ts
@@ -11,9 +11,9 @@ import { readdir, readFile } from "fs-extra";
 import { load } from "js-yaml";
 import { dirname, join } from "path";
 import { fetchExternalApiQueries } from "../../../../src/data-extensions-editor/queries";
-import * as log from "../../../../src/common/vscode/log";
+import * as vscodeLog from "../../../../src/common/vscode/logging";
 import { RedactableError } from "../../../../src/pure/errors";
-import { showAndLogExceptionWithTelemetry } from "../../../../src/common/vscode/log";
+import { showAndLogExceptionWithTelemetry } from "../../../../src/common/vscode/logging";
 
 function createMockUri(path = "/a/b/c/foo"): Uri {
   return {
@@ -140,7 +140,7 @@ describe("readQueryResults", () => {
 
   beforeEach(() => {
     showAndLogExceptionWithTelemetrySpy = jest.spyOn(
-      log,
+      vscodeLog,
       "showAndLogExceptionWithTelemetry",
     );
   });
